% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_auto_swap.R
\name{oku_optimal_swap}
\alias{oku_optimal_swap}
\title{Automated Optimal Swap with Oku API}
\usage{
oku_optimal_swap(
  pool_address,
  chain,
  min_range,
  max_range,
  x_init = 0,
  y_init = 0,
  usd_init = 0
)
}
\arguments{
\item{pool_address}{Character. The contract address of the liquidity pool.}

\item{chain}{Character. The blockchain name as used by Oku (e.g., "gnosis", "arbitrum", "optimism").}

\item{min_range}{Numeric. Lower bound of the range (expressed as amount of token1 per token0).}

\item{max_range}{Numeric. Upper bound of the range (expressed as amount of token1 per token0).}

\item{x_init}{Numeric. Initial amount of token0 (X) in the wallet.}

\item{y_init}{Numeric. Initial amount of token1 (Y) in the wallet.}

\item{usd_init}{Numeric. Initial amount of USD to invest.}
}
\description{
Fetches the real-time USD prices of the tokens in a specified
Uniswap V3 pool using the Oku (Icarus) API, and then calculates the optimal swap.
}
\examples{
oku_optimal_swap(
 pool_address = "0xe8a249626d3f3b876b887c30a3355513cb3fa9e4",
 chain = "gnosis",
 min_range = 2000,
 max_range = 2500,
 x_init = 0,
 y_init = 893.91,
 usd_init = 0)

}
