% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{price_to_ticks}
\alias{price_to_ticks}
\title{Get bounding prices based on fee tier}
\usage{
price_to_ticks(p, fee)
}
\arguments{
\item{p}{Numeric. The current or target price.}

\item{fee}{Numeric or Character. The pool fee tier
(accepted values: 100, 500, 3000, 10000).}
}
\value{
A numeric vector of length 2 containing the lower bound price
  and the upper bound price.
}
\description{
Given a target price and a pool fee tier, this function determines
the applicable tick spacing and returns the prices corresponding
to the valid ticks immediately below and above the given price.
}
\examples{
price_to_ticks(p = 1500, fee = 500)
price_to_ticks(p = 1500, fee = 3000)
}
