% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tick_to_price}
\alias{tick_to_price}
\title{Convert tick to price}
\usage{
tick_to_price(i)
}
\arguments{
\item{i}{Integer or Numeric. The tick index.}
}
\value{
A numeric value representing the price.
}
\description{
Calculates the exact price corresponding to a specific tick index.
}
\examples{
tick_to_price(73145)
}
