% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_optimal_swap.R
\name{lp_optimal_swap}
\alias{lp_optimal_swap}
\title{Optimal Swap Calculation for Uniswap V3}
\usage{
lp_optimal_swap(
  price_x_usd,
  price_y_usd,
  min_range,
  max_range,
  x_init = 0,
  y_init = 0,
  usd_init = 0
)
}
\arguments{
\item{price_x_usd}{Numeric. Price of token X in USD.}

\item{price_y_usd}{Numeric. Price of token Y in USD.}

\item{min_range}{Numeric. Lower bound of the range (expressed as amount of Y per 1 X).}

\item{max_range}{Numeric. Upper bound of the range (expressed as amount of Y per 1 X).}

\item{x_init}{Numeric. Initial amount of token X in the wallet.}

\item{y_init}{Numeric. Initial amount of token Y in the wallet.}

\item{usd_init}{Numeric. Initial amount of USD to invest (Optional).}
}
\value{
A list containing the details of the action to be performed:
\itemize{
  \item \code{action} : The type of swap to execute (e.g., "Swap X for Y").
  \item \code{amount} : The amount of tokens to swap.
  \item \code{amount_usd} : The USD value of the swap.
  \item \code{token_in} : The token to sell.
  \item \code{token_out} : The token to buy.
}
}
\description{
Calculates the exact amount of tokens to swap to reach a perfect
asset ratio before adding liquidity to a Uniswap V3 pool (or similar)
based on a defined price range.
}
\examples{
lp_optimal_swap(
  price_x_usd = 3553,
  price_y_usd = 1.1656,
  min_range = 2800,
  max_range = 3200,
  x_init = 1,
  y_init = 2000,
  usd_init = 0)

lp_optimal_swap(
  price_x_usd = 3553,
  price_y_usd = 1.1656,
  min_range = 2800,
  max_range = 3200,
  x_init = 0,
  y_init = 0,
  usd_init = 5884.20)

}
